% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.vsflcm.R
\name{cv.vsflcm}
\alias{cv.vsflcm}
\title{Cross validation to choose lambda in vsflcm}
\usage{
cv.vsflcm(formula, data = NULL, id.time = NULL, id.sub = NULL,
  t.min = NULL, t.max = NULL, K = 5, spline.fun = "B-spline",
  lambdas = 10^seq(-2, 2, 1), K0 = 2, lam.nuc = 10,
  method.obj = c("nuclear", "nonconvex"), delta = 0.1,
  method.optim = c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent"),
  times = 1, fpc.on = TRUE, seed.split = 1, num.fold = 5)
}
\arguments{
\item{formula}{formula for the regression. should have form \code{ y ~ V1 + V2 + ... + Vk}. Note: don't contain \code{id.time}}

\item{data}{data frame}

\item{id.time}{the variable that represents time}

\item{id.sub}{variable giving subject ID}

\item{t.min}{minimum value to be evaluated on the time domain. if `NULL`, taken to be minium observed value.}

\item{t.max}{maximum value to be evaluated on the time domain. if `NULL`, taken to be minium observed value.}

\item{K}{number of spline basis functions for coefficients and fpc}

\item{spline.fun}{spline basis functions. If not default, it should be a function taking a vector(corresponding to time) belonging to [0,1] as input and returning a \code{length(vector)*K} spline matrix where each row corresponds to a time point in the vector and each column corresponds to a spline basis function}

\item{lambdas}{tuning parameter vector for \code{lambda} to choose from}

\item{K0}{number of FPCs}

\item{lam.nuc}{Only Useful when \code{method.obj}='nuclear'. penalty factor for the nuclear norm}

\item{method.obj}{the methodology chosen. Should be 'nuclear' or 'nonconvex'. Typically, we recommend 'nuclear'}

\item{delta}{Only useful when \code{method.obj}='nonconvex'. When the square of the norm of the coefficient is smaller than \code{delta}, L1 penalty will smoothly turned into L2 penalty}

\item{method.optim}{the optimization method to be used. For details and other options, please refer to \code{optim}}

\item{times}{Only useful when \code{method.obj}='nonconvex'. the algorithm randomly chooses \code{times} initial points to optimize and select the best}

\item{fpc.on}{Logical. Whether fpc will be used}

\item{seed.split}{seed to randomly split the data set into training set and test set}

\item{num.fold}{\code{num.fold}-fold cross validation}
}
\description{
Cross validation to choose lambda in vsflcm
}
\author{
Hongming Pu \email{phmhappier@163.com}
}
