% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myInt.R
\name{myInt}
\alias{myInt}
\title{Internal function used to compute a special case of integral faster}
\usage{
myInt(fun, K, tol = 1e-06, start = 0, end = 1)
}
\arguments{
\item{fun}{basic function}

\item{K}{number of spline basis functions}

\item{tol}{error tolerance of the integral}

\item{start}{start point of the interval}

\item{end}{end point of the interval}
}
\description{
Internal function used to compute a special case of integral faster
}
\author{
Hongming Pu \email{phmhappier@163.com}
}
