% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.flcmgl.R
\name{predict.flcmgl}
\alias{predict.flcmgl}
\title{predict for a new dataset based on a fitted concurrent functional linear model}
\usage{
\method{predict}{flcmgl}(obj, data, sub.reg = TRUE)
}
\arguments{
\item{obj}{object returned from \code{\link{vsflcm}}}

\item{data}{dataset which fitted values should be computed for}

\item{sub.reg}{Logical, whether the subjects should be recognized based on the knowledge from the training data}
}
\value{
a vector of fitted values
}
\description{
predict for a new dataset based on a fitted concurrent functional linear model
}
\author{
Hongming Pu \email{phmhappier@163.com}
}
