% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanFit.R
\name{stanFit}
\alias{stanFit}
\title{This function standarizes the predictors}
\usage{
stanFit(data)
}
\arguments{
\item{data}{data to be standarized}
}
\description{
This function standarizes the predictors
}
\author{
Hongming Pu \email{phmhappier@163.com}
}
