% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.par.vsflcm.R
\name{tune.par.vsflcm}
\alias{tune.par.vsflcm}
\title{tune key parameters(K,lambda,lam.nuc,lam.smo) using cross validation in vsflcm, if you don't want to tune the parameter by yourself, this is for you.}
\usage{
tune.par.vsflcm(formula, data = NULL, id.time = NULL, id.sub = NULL,
  t.min = NULL, t.max = NULL, seed.split = 4, num.fold = 3,
  thre = 0.9, fpc.on = TRUE, intercept = TRUE, K = NULL,
  lambdas = 10^seq(-2, 2, 1), lam.smos = c(0, 10^seq(-2, 2, 1)),
  lam.nucs = c(0, 10^seq(-2, 2, 1)), K0 = 2, spline.fun = "B-spline",
  method.obj = "nuclear", K.max = 10, delta = 0.1,
  method.optim = "BFGS", times = 1, spline.fun2d = NULL, maxit = 200)
}
\arguments{
\item{formula}{formula for the regression. should have form \code{ y ~ V1 + V2 + ... + Vk}. Note: don't contain \code{id.time}}

\item{data}{data frame}

\item{id.time}{the variable that represents time}

\item{id.sub}{variable giving subject ID}

\item{t.min}{minimum value to be evaluated on the time domain. if `NULL`, taken to be minium observed value.}

\item{t.max}{maximum value to be evaluated on the time domain. if `NULL`, taken to be minium observed value.}

\item{seed.split}{seed to randomly split the data set into training set and test set}

\item{num.fold}{\code{num.fold}-fold cross validation}

\item{thre}{when automaticly chosing K, K will be chosen as the smallest one which ensures that the model can explain \code{thre} of the total variance.}

\item{fpc.on}{Logical. Whether fpc will be used}

\item{intercept}{logical, whether intercept is used}

\item{K}{number of spline basis functions for coefficients and fpc. If default this function will automaticly choose K. Otherwise you need to provide a value.}

\item{lambdas}{tuning parameter vector for \code{lambda} to choose from}

\item{lam.smos}{tuning parameter vector for \code{lam.smo} to choose from}

\item{lam.nucs}{tuning parameter vector for \code{lam.nuc} to choose from}

\item{K0}{number of FPCs}

\item{spline.fun}{spline basis functions. If not default, it should be a function taking a vector(corresponding to time) belonging to [0,1] as input and returning a \code{length(vector)*K} spline matrix where each row corresponds to a time point in the vector and each column corresponds to a spline basis function}

\item{method.obj}{the methodology chosen. Should be 'nuclear' or 'nonconvex'. Typically, we recommend 'nuclear'}

\item{K.max}{max \code{K} can be chosen automaticly, only useful when \code{K} is NULL}

\item{delta}{Only useful when \code{method.obj}='nonconvex'. When the square of the norm of the coefficient is smaller than \code{delta}, L1 penalty will smoothly turned into L2 penalty}

\item{method.optim}{the optimization method to be used. For details and other options, please refer to \code{optim}}

\item{times}{Only useful when \code{method.obj}='nonconvex'. the algorithm randomly chooses \code{times} initial points to optimize and select the best}

\item{spline.fun2d}{second derivative of \code{spline.fun}, if \code{spline.fun} is default then the second derivative will be computed automaticly and you don't need to provide values for \code{spline.fun2d}. Otherwise you need to provide a function with input and output dimensions in accordance with \code{spline.fun} if you want to model the smoothness.}

\item{maxit}{\code{maxit} of \code{optim}}
}
\description{
tune key parameters(K,lambda,lam.nuc,lam.smo) using cross validation in vsflcm, if you don't want to tune the parameter by yourself, this is for you.
}
\author{
Hongming Pu \email{phmhappier@163.com}
}
