% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsflcm.R
\name{vsflcm}
\alias{vsflcm}
\title{Implements group lasso with fpc for the linear functional concurrent model}
\usage{
vsflcm(formula, data = NULL, id.time = NULL, intercept = TRUE,
  id.sub = NULL, t.min = NULL, t.max = NULL, K = 5,
  spline.fun = "B-spline", lambda = 0.5, K0 = 2, lam.nuc = 10,
  method.obj = c("nuclear", "nonconvex"), delta = 0.1,
  method.optim = c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent"),
  times = 1, fpc.on = TRUE, spline.fun2d = NULL, lam.smo = 0,
  maxit = 100)
}
\arguments{
\item{formula}{formula for the regression. should have form \code{ y ~ V1 + V2 + ... + Vk}. Note: don't contain \code{id.time}}

\item{data}{data frame}

\item{id.time}{the variable that represents time}

\item{intercept}{logical, whether use intercept function}

\item{id.sub}{variable giving subject ID}

\item{t.min}{minimum value to be evaluated on the time domain. if `NULL`, taken to be minium observed value.}

\item{t.max}{maximum value to be evaluated on the time domain. if `NULL`, taken to be minium observed value.}

\item{K}{number of spline basis functions for coefficients and fpc}

\item{spline.fun}{spline basis functions. If not default, it should be a function taking a vector(corresponding to time) belonging to [0,1] as input and returning a \code{length(vector)*K} spline matrix where each row corresponds to a time point in the vector and each column corresponds to a spline basis function}

\item{lambda}{Constant to multiply the L1 penalty term}

\item{K0}{number of FPCs}

\item{lam.nuc}{Only Useful when \code{method.obj}='nuclear'. penalty factor for the nuclear norm}

\item{method.obj}{the methodology chosen. Should be 'nuclear' or 'nonconvex'. Typically, we recommend 'nuclear'}

\item{delta}{Only useful when \code{method.obj}='nonconvex'. When the square of the norm of the coefficient is smaller than \code{delta}, L1 penalty will smoothly turned into L2 penalty}

\item{method.optim}{the optimization method to be used. For details and other options, please refer to \code{optim}}

\item{times}{Only useful when \code{method.obj}='nonconvex'. the algorithm randomly chooses \code{times} initial points to optimize and select the best}

\item{fpc.on}{Logical. Whether fpc will be used}

\item{spline.fun2d}{second derivative of \code{spline.fun}, if \code{spline.fun} is default then the second derivative will be computed automaticly and you don't need to provide values for \code{spline.fun2d}. Otherwise you need to provide a function with input and output dimensions in accordance with \code{spline.fun} if you want to model the smoothness.}

\item{lam.smo}{penalty factor for the second derivative of basis funtion. Only useful when \code{spline.fun} is default or \code{spline.fun2d} is not NULL}

\item{maxit}{\code{maxit} of \code{optim}}
}
\description{
Implements group lasso with fpc for the linear functional concurrent model
}
\author{
Hongming Pu \email{phmhappier@163.com}
}
