% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsflcm.R
\name{vsflcm}
\alias{vsflcm}
\title{vb_concurrent}
\usage{
vsflcm(formula, data = NULL, id.time = NULL, id.sub = NULL,
  t.min = NULL, t.max = NULL, K = 5, spline.fun = "B-spline",
  lambda = 0.5, K0 = 2, delta = 1e-05, method.optim = "BFGS",
  times = 1, fpc.on = TRUE)
}
\arguments{
\item{formula}{formula for the regression. should have form \code{ y ~ V1 + V2 + ... + Vk}. Note: don't contain \code{id.time}}

\item{data}{data frame}

\item{id.time}{the variable that represents time}

\item{id.sub}{variable giving subject ID}

\item{t.min}{minimum value to be evaluated on the time domain. if `NULL`, taken to be minium observed value.}

\item{t.max}{maximum value to be evaluated on the time domain. if `NULL`, taken to be minium observed value.}

\item{K}{number of spline basis functions for coefficients and fpc}

\item{spline.fun}{spline basis functions. If not default, it should be a function taking a vector(corresponding to time) belonging to [0,1] as input and returning a \code{length(vector)*K} spline matrix where each row corresponds to a time point in the vector and each column corresponds to a spline basis function}

\item{lambda}{Constant multiplying the L1 penalty term}

\item{K0}{number of FPCs}

\item{delta}{When the square of the norm of the coefficient is smaller than \code{delta}, L1 penalty will smoothly turned into L2 penalty}

\item{method.optim}{the optimization method to be used. For details and other options, please refer to \code{optim}}

\item{times}{the algorithm randomly chooses \code{times} initial points to optimize and select the best}

\item{fpc.on}{Logical. Whether fpc will be used}
}
\description{
Implements group lasso with fpc for the linear functional concurrent model
}
\author{
Hongming Pu \email{phmhappier@163.com}
}
